Removed singleton structure to allow multiple connections.
Instead of using the CineastWrapper singleton, a CineastClient can be created for a specific configuration.
The Segment and Object registries have also been refactored into a single instantiated class MultimediaRegistry.
As a consequence of the changes, segments and objects must retain references to which CineastClient they have been created from.